organizations datacube documentation

organizations datacube documentation

Usage

organizations

Format

The organizations datacube is a list that contains the following 3 datasets: CHATHAM, IHEID, GHS. For more information and references to each of the datasets used, please use the data_source() and data_contrast() functions.

CHATHAM:

A dataset with 203 observations and the following 7 variables: organizationID, Organization, Begin, City, State, Type, Health_as_primary_intent.

IHEID:

A dataset with 124 observations and the following 5 variables: organizationID, Organization, Begin, Areas, City.

GHS:

A dataset with 203 observations and the following 8 variables: organizationID, Organization, Begin, City, State, URL, Type, Health as primary intent?.

Source

CHATHAM: S. J. Hoffman, C. B. Cole, and M. Pearcey. In: Mapping global health architecture to inform thefuture. Chatham House London, 2015.

IHEID: In: Directory of Global Health Actors. 2021.

GHS: S. J. Hoffman and C. B. Cole. “Defining the global health system and systematically mapping itsnetwork of actors”. In: Global Health 14.38 (2018).

Details

#> $CHATHAM
#> -------------------------------------------------------------------------------------
#> |      Column Name       |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> -------------------------------------------------------------------------------------
#> |     organizationID     |  character  |             203|     0     |       0       |
#> |      Organization      |  character  |             203|     0     |       0       |
#> |         Begin          |    mdate    |             203|     0     |       0       |
#> |          City          |  character  |             203|     0     |       0       |
#> |         State          |  character  |             203|     0     |       0       |
#> |          Type          |   numeric   |             203|     0     |       0       |
#> |Health_as_primary_intent|  character  |             203|     0     |       0       |
#> -------------------------------------------------------------------------------------
#> 
#> 
#> $IHEID
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |organizationID |  character  |             124|    16     |     12.9      |
#> | Organization  |  character  |             124|     0     |       0       |
#> |     Begin     |    mdate    |             124|    44     |     35.48     |
#> |     Areas     |  character  |             124|     3     |     2.42      |
#> |     City      |  character  |             124|     3     |     2.42      |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $GHS
#> --------------------------------------------------------------------------------------
#> |       Column Name       |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> --------------------------------------------------------------------------------------
#> |     organizationID      |  character  |             203|     0     |       0       |
#> |      Organization       |  character  |             203|     0     |       0       |
#> |          Begin          |    mdate    |             203|     0     |       0       |
#> |          City           |  character  |             203|     0     |       0       |
#> |          State          |  character  |             203|     0     |       0       |
#> |           URL           |  character  |             203|     0     |       0       |
#> |          Type           |  character  |             203|     0     |       0       |
#> |Health as primary intent?|  character  |             203|     0     |       0       |
#> --------------------------------------------------------------------------------------

URL

CHATHAM: https://www.chathamhouse.org/sites/default/files/field/field_document/20150120GlobalHealthArchitectureHoffmanColePearcey.pdf

IHEID: https://www.graduateinstitute.ch/GHC-directories

GHS: https://doi.org/10.1186/s12992-018-0340-2

Mapping

CHATHAM: Variable Mapping

from	to

Code	organizationID

Actors	Organization

HQ_states	State

HQ_city	City

Year_inception	Begin

IHEID: Variable Mapping

from	to

Abbreviation	organizationID

Name	Organization

Start	Begin

'HQ/Offices'	City

GHS: Variable Mapping

from	to

Code	organizationID

Actor	Organization

'Year of Inception'	Begin

'Headquarters Location'	City