agreements datacube documentation

agreements datacube documentation

Usage

agreements

Format

The agreements datacube is a list that contains the following 3 datasets: WHO, GHHR, HUGGO. For more information and references to each of the datasets used, please use the manydata::call_sources() and manydata::compare_dimensions() functions.

WHO:

A dataset with 108 observations and the following 8 variables: manyID, treatyID, Title, Begin, Organisation, Topic, Lineage, whoID.

GHHR:

A dataset with 149 observations and the following 8 variables: manyID, treatyID, Title, Begin, Region, LegalStatus, Lineage, ghhrID.

HUGGO:

A dataset with 252 observations and the following 13 variables: manyID, treatyID, Title, Begin, Signature, Force, End, Formal, TreatyText, url, Source, Comments, Topic.

Source

WHO: D. Bhugra, S. Pathare, R. Joshi, et al. “A review of mental health policies from Commonwealthcountries”. International Journal of Social Psychiatry 64.1 (2018), pp. 3-8.

GHHR: B. M. Meier, O. A. Cabrera, A. Ayala, et al. “Bridging international law and rights-based litigation:mapping health-related rights through the development of the global health and human rights database”. Health & Human Rights. 14 (2012), p. 20.

HUGGO: J. Hollway. Health agreements for manydata. 2021.

Details

#> $WHO
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             108|     0     |       0       |
#> |   treatyID    |  character  |             108|     0     |       0       |
#> |     Title     |  character  |             108|     0     |       0       |
#> |     Begin     |    mdate    |             108|     0     |       0       |
#> | Organisation  |  character  |             108|     0     |       0       |
#> |     Topic     |  character  |             108|     0     |       0       |
#> |    Lineage    |  character  |             108|     0     |       0       |
#> |     whoID     |   integer   |             108|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $GHHR
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             149|     0     |       0       |
#> |   treatyID    |  character  |             149|     0     |       0       |
#> |     Title     |  character  |             149|     0     |       0       |
#> |     Begin     |    mdate    |             149|     3     |     2.01      |
#> |    Region     |  character  |             149|     0     |       0       |
#> |  LegalStatus  |  character  |             149|     0     |       0       |
#> |    Lineage    |  character  |             149|    52     |     34.9      |
#> |    ghhrID     |  character  |             149|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             252|     0     |       0       |
#> |   treatyID    |  character  |             252|     0     |       0       |
#> |     Title     |  character  |             252|     0     |       0       |
#> |     Begin     |    mdate    |             252|     0     |       0       |
#> |   Signature   |    mdate    |             252|     0     |       0       |
#> |     Force     |    mdate    |             252|    179    |     71.03     |
#> |      End      |    mdate    |             252|     0     |       0       |
#> |    Formal     |  character  |             252|     0     |       0       |
#> |  TreatyText   |  character  |             252|     0     |       0       |
#> |      url      |  character  |             252|    123    |     48.81     |
#> |    Source     |  character  |             252|    118    |     46.83     |
#> |   Comments    |  character  |             252|    229    |     90.87     |
#> |     Topic     |  character  |             252|     0     |       0       |
#> ----------------------------------------------------------------------------

URL

WHO: https://www.mindbank.info/collection/un_who_resolutions/all?page=all

GHHR: https://www.globalhealthrights.org/instruments/instrument-region/

HUGGO: Hand-coded data by the GGO team

Mapping

WHO: Variable Mapping

from	to

NA	NA

GHHR: Variable Mapping

from	to

Year of adoption	Begin

HUGGO: Variable Mapping

from	to

NA	NA