organizations database documentation

organizations database documentation

Usage

organizations

Format

The organizations database is a list that contains the following 3 datasets: CHATHAM, IHEID, GHS. If using the data in combination with manystates data, please note that there may be overlaps in three-letter organizationID and three-letter stateID. For more information and references to each of the datasets used, please use the data_source() and data_contrast() functions.

CHATHAM:

A dataset with 203 observations and the following 7 variables: organizationID, Organization, Beg, City, State, Type, Health_as_primary_intent.

IHEID:

A dataset with 124 observations and the following 5 variables: organizationID, Organization, Beg, Areas, City.

GHS:

A dataset with 203 observations and the following 8 variables: organizationID, Organization, Beg, City, State, URL, Type, Health as primary intent?.

Details

#> $CHATHAM
#> -------------------------------------------------------------------------------------
#> |      Column Name       |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> -------------------------------------------------------------------------------------
#> |     organizationID     |  character  |             203|     0     |       0       |
#> |      Organization      |  character  |             203|     0     |       0       |
#> |          Beg           |    mdate    |             203|     0     |       0       |
#> |          City          |  character  |             203|     0     |       0       |
#> |         State          |  character  |             203|     0     |       0       |
#> |          Type          |   numeric   |             203|     0     |       0       |
#> |Health_as_primary_intent|  character  |             203|     0     |       0       |
#> -------------------------------------------------------------------------------------
#> 
#> 
#> $IHEID
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |organizationID |  character  |             124|    16     |     12.9      |
#> | Organization  |  character  |             124|     0     |       0       |
#> |      Beg      |    mdate    |             124|    44     |     35.48     |
#> |     Areas     |  character  |             124|     3     |     2.42      |
#> |     City      |  character  |             124|     3     |     2.42      |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $GHS
#> --------------------------------------------------------------------------------------
#> |       Column Name       |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> --------------------------------------------------------------------------------------
#> |     organizationID      |  character  |             203|     0     |       0       |
#> |      Organization       |  character  |             203|     0     |       0       |
#> |           Beg           |    mdate    |             203|     0     |       0       |
#> |          City           |  character  |             203|     0     |       0       |
#> |          State          |  character  |             203|     0     |       0       |
#> |           URL           |  character  |             203|     0     |       0       |
#> |          Type           |  character  |             203|     0     |       0       |
#> |Health as primary intent?|  character  |             203|     0     |       0       |
#> --------------------------------------------------------------------------------------